+++ bzt/modules/jmeter.py
@@ -35,7 +35,7 @@ from cssselect import GenericTranslator
 from bzt.engine import ScenarioExecutor, Scenario, FileLister
 from bzt.jmx import JMX
 from bzt.modules.aggregator import ConsolidatingAggregator, ResultsReader, DataPoint, KPISet
-from bzt.modules.console import WidgetProvider, SidebarWidget
+from bzt.modules.console import WidgetProvider, ExecutorWidget
 from bzt.modules.provisioning import Local
 from bzt.six import iteritems, string_types, StringIO, etree, binary_type
 from bzt.six import parse, request as http_request
@@ -53,11 +53,11 @@ class JMeterExecutor(ScenarioExecutor, WidgetProvider, FileLister):
     :type properties_file: str
     :type sys_properties_file: str
     """
-    MIRRORS_SOURCE = "https://archive.apache.org/dist/jmeter/binaries/"
+    MIRRORS_SOURCE = "https://jmeter.apache.org/download_jmeter.cgi"
     JMETER_DOWNLOAD_LINK = "https://archive.apache.org/dist/jmeter/binaries/apache-jmeter-{version}.zip"
-    PLUGINS_MANAGER = 'http://search.maven.org/remotecontent?filepath=' \
+    PLUGINS_MANAGER = 'https://search.maven.org/remotecontent?filepath=' \
                       'kg/apc/jmeter-plugins-manager/0.8/jmeter-plugins-manager-0.8.jar'
-    CMDRUNNER = 'http://search.maven.org/remotecontent?filepath=kg/apc/cmdrunner/2.0/cmdrunner-2.0.jar'
+    CMDRUNNER = 'https://search.maven.org/remotecontent?filepath=kg/apc/cmdrunner/2.0/cmdrunner-2.0.jar'
     JMETER_VER = "3.0"
     UDP_PORT_NUMBER = None
 
@@ -193,9 +193,6 @@ class JMeterExecutor(ScenarioExecutor, WidgetProvider, FileLister):
         :return: bool
         :raise RuntimeWarning:
         """
-        if self.widget:
-            self.widget.update()
-
         self.retcode = self.process.poll()
         if self.retcode is not None:
             if self.retcode != 0:
@@ -617,7 +614,7 @@ class JMeterExecutor(ScenarioExecutor, WidgetProvider, FileLister):
         """
         if not self.widget:
             label = "%s" % self
-            self.widget = SidebarWidget(self, "JMeter: " + label.split('/')[1])
+            self.widget = ExecutorWidget(self, "JMeter: " + label.split('/')[1])
         return self.widget
 
     def __modify_resources_paths_in_jmx(self, jmx, file_list):
@@ -770,7 +767,7 @@ class JMeterExecutor(ScenarioExecutor, WidgetProvider, FileLister):
         jmeter_path = self.settings.get("path", "~/.bzt/jmeter-taurus/")
         jmeter_path = get_full_path(jmeter_path)
         jmeter_version = self.settings.get("version", JMeterExecutor.JMETER_VER)
-        download_link = self.settings.get("download-link", JMeterExecutor.JMETER_DOWNLOAD_LINK)
+        download_link = self.settings.get("download-link", None)
         plugins = self.settings.get("plugins", [])
         proxy = self.engine.config.get('settings').get('proxy')
         tool = JMeter(jmeter_path, self.log, jmeter_version, download_link, plugins, proxy)
@@ -995,8 +992,6 @@ class JTLErrorsReader(object):
     def read_file(self):
         """
         Read the next part of the file
-
-        :return:
         """
 
         if self.failed_processing:
@@ -1042,9 +1037,6 @@ class JTLErrorsReader(object):
     def get_data(self, max_ts):
         """
         Get accumulated errors data up to specified timestamp
-
-        :param max_ts:
-        :return:
         """
         result = BetterDict()
         for t_stamp in sorted(self.buffer.keys()):
@@ -1138,6 +1130,8 @@ class JTLErrorsReader(object):
     def __get_failed_assertion(self, element):
         """
         Returns first failed assertion, or None
+
+        :rtype lxml.etree.Element
         """
         assertions = [elem for elem in element.iterchildren() if elem.tag == "assertionResult"]
         for assertion in assertions:
@@ -1521,10 +1515,10 @@ class JMeter(RequiredTool):
     """
 
     def __init__(self, tool_path, parent_logger, jmeter_version, jmeter_download_link, plugins, proxy):
-        super(JMeter, self).__init__("JMeter", tool_path)
+        super(JMeter, self).__init__("JMeter", tool_path, jmeter_download_link)
         self.log = parent_logger.getChild(self.__class__.__name__)
         self.version = jmeter_version
-        self.mirror_manager = JMeterMirrorsManager(self.log, self.version, jmeter_download_link)
+        self.mirror_manager = JMeterMirrorsManager(self.log, self.version)
         self.plugins = plugins
         self.proxy_settings = proxy
 
@@ -1552,11 +1546,17 @@ class JMeter(RequiredTool):
             return False
 
     def __install_jmeter(self, dest):
-        with super(JMeter, self).install_with_mirrors(dest, ".zip") as jmeter_dist:
+        if self.download_link:
+            jmeter_dist = super(JMeter, self).install_with_link(dest, ".zip")
+        else:
+            jmeter_dist = super(JMeter, self).install_with_mirrors(dest, ".zip")
+
+        try:
             self.log.info("Unzipping %s to %s", jmeter_dist.name, dest)
             unzip(jmeter_dist.name, dest, 'apache-jmeter-%s' % self.version)
-
-        os.remove(jmeter_dist.name)
+        finally:
+            jmeter_dist.close()
+            os.remove(jmeter_dist.name)
 
         # set exec permissions
         os.chmod(os.path.join(dest, 'bin', 'jmeter'), 0o755)
@@ -1568,14 +1568,14 @@ class JMeter(RequiredTool):
     def __download_additions(self, tools):
         downloader = http_request.FancyURLopener()
         with ProgressBarContext() as pbar:
-            try:
-                for tool in tools:
-                    _file = os.path.basename(tool[0])
+            for tool in tools:
+                _file = os.path.basename(tool[0])
+                try:
                     self.log.info("Downloading %s from %s", _file, tool[0])
                     downloader.retrieve(tool[0], tool[1], pbar.download_callback)
-            except BaseException:
-                self.log.error("Error while downloading %s", _file)
-                raise
+                except BaseException:
+                    self.log.error("Error while downloading %s", _file)
+                    raise
 
     def __install_plugins_manager(self, plugins_manager_path):
         installer = "org.jmeterplugins.repository.PluginManagerCMDInstaller"
@@ -1596,28 +1596,29 @@ class JMeter(RequiredTool):
         try:
             # prepare proxy settings
             if self.proxy_settings and self.proxy_settings.get('address'):
+                env = BetterDict()
+                env.merge(dict(os.environ))
+                jvm_args = env.get('JVM_ARGS', '')
+
                 proxy_url = parse.urlsplit(self.proxy_settings.get("address"))
                 self.log.debug("Using proxy settings: %s", proxy_url)
                 host = proxy_url.hostname
                 port = proxy_url.port
                 if not port:
                     port = 80
+
+                jvm_args += ' -Dhttp.proxyHost=%s -Dhttp.proxyPort=%s' % (host, port)  # TODO: remove it after pmgr 0.9
+                jvm_args += ' -Dhttps.proxyHost=%s -Dhttps.proxyPort=%s' % (host, port)
+
                 username = self.proxy_settings.get('username')
                 password = self.proxy_settings.get('password')
-                host_to_jvm = '-Dhttp.proxyHost=%s -Dhttp.proxyPort=%s' % (host, port)
-                if username and password:
-                    auth_to_jvm = '-Dhttp.proxyUsername="%s" -Dhttp.proxyPassword="%s"' % (username, password)
-                else:
-                    auth_to_jvm = ''
 
-                env = BetterDict()
-                env.merge(dict(os.environ))
-                jvm_args = env.get('JVM_ARGS', '')
-                if jvm_args:
-                    jvm_args += ' '
-                if auth_to_jvm:
-                    auth_to_jvm += ' '
-                env['JVM_ARGS'] = env.get('JVM_ARGS', '') + host_to_jvm + auth_to_jvm
+                if username and password:
+                    # property names correspond to
+                    # https://github.com/apache/jmeter/blob/trunk/src/core/org/apache/jmeter/JMeter.java#L110
+                    jvm_args += ' -Dhttp.proxyUser="%s" -Dhttp.proxyPass="%s"' % (username, password)
+
+                env['JVM_ARGS'] = jvm_args
 
             proc = shell_exec(cmd)
             out, err = proc.communicate()
@@ -1672,7 +1673,7 @@ class JarCleaner(object):
         for jar_lib_obj in jar_libs:
             similar_packages = [lib for lib in jar_libs if lib.lib_name == jar_lib_obj.lib_name]
             if len(similar_packages) > 1:
-                right_version = max(similar_packages, key=lambda lib: LooseVersion(lib.version))
+                right_version = max(similar_packages, key=lambda l: LooseVersion(l.version))
                 similar_packages.remove(right_version)
                 duplicated_libraries.update(similar_packages)
 
@@ -1682,28 +1683,26 @@ class JarCleaner(object):
 
 
 class JMeterMirrorsManager(MirrorsManager):
-    def __init__(self, parent_logger, jmeter_version, download_link):
+    def __init__(self, parent_logger, jmeter_version):
         self.jmeter_version = str(jmeter_version)
-        self.download_link = download_link
         super(JMeterMirrorsManager, self).__init__(JMeterExecutor.MIRRORS_SOURCE, parent_logger)
 
     def _parse_mirrors(self):
         links = []
         if self.page_source is not None:
             self.log.debug('Parsing mirrors...')
-            href_search_pattern = re.compile(r'<a href="apache\-jmeter\-([^"]*?)\.zip">')
-            jmeter_versions = href_search_pattern.findall(self.page_source)
-
-            if self.jmeter_version in jmeter_versions:
-                archive_name = "apache-jmeter-{version}.zip".format(version=self.jmeter_version)
-                links.append(self.base_link + archive_name)
-            else:
-                self.log.warning("Can't find link for JMeter %s, will use download-link", self.jmeter_version)
-        default_link = self.download_link.format(version=self.jmeter_version)
-        if default_link not in links:
-            links.append(default_link)
+            select_search_pattern = re.compile(r'<select name="Preferred">.*?</select>', re.MULTILINE | re.DOTALL)
+            option_search_pattern = re.compile(r'<option value=".*?">')
+            select_element = select_search_pattern.findall(self.page_source)
+            if select_element:
+                option_elements = option_search_pattern.findall(select_element[0])
+                link_tail = "/jmeter/binaries/apache-jmeter-{version}.zip".format(version=self.jmeter_version)
+                links = [link.strip('<option value="').strip('">') + link_tail for link in option_elements]
+        links.append(JMeterExecutor.JMETER_DOWNLOAD_LINK.format(version=self.jmeter_version))
         self.log.debug('Total mirrors: %d', len(links))
-        return links
+        # place HTTPS links first, preserving the order of HTTP links
+        sorted_links = sorted(links, key=lambda l: l.startswith("https"), reverse=True)
+        return sorted_links
 
 
 class Request(object):